requirements.txt
numpy>=1.24.0
pyrtlsdr>=0.2.93
pyyaml>=6.0